requirements.txt
python-dotenv
SQLAlchemy
psycopg2-binary
pandas
requests
